Added password validation, response serializer, moved investor and startup models to their app
