+++ forum/users/models.py
@@ -42,39 +42,5 @@ class User(AbstractBaseUser, PermissionsMixin):
         verbose_name_plural = 'users'
         db_table = 'users'
 
-class Investor(models.Model):
-    investor_id = models.AutoField(primary_key=True)  
-    user_id = models.OneToOneField(
-        User,
-        on_delete=models.CASCADE,
-        db_column='user_id'
-    )
-    company_name = models.CharField(max_length=255)
-    industry = models.CharField(max_length=255, blank=True, null=True)
-    created_at = models.DateField(auto_now_add=True)
-    address = models.CharField(max_length=255)
-    website = models.URLField(blank=True, null=True)
 
 
-    class Meta:
-        verbose_name = 'investor'
-        verbose_name_plural = 'investors'
-        db_table = 'investors'
-
-class Startup(models.Model):
-    startup_id = models.AutoField(primary_key=True)
-    user_id = models.OneToOneField(
-        User,
-        on_delete=models.CASCADE,
-        db_column='user_id'
-    )
-    title = models.CharField(max_length=255)
-    description = models.TextField()
-    industry = models.CharField(max_length=255, blank=True, null=True)
-    website = models.CharField(max_length=255, blank=True, null=True)
-    address = models.CharField(max_length=255, blank=True, null=True)
-
-    class Meta:
-        verbose_name = 'startup'
-        verbose_name_plural = 'startups'
-        db_table = 'startups'
